% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquacrop_morris.R
\name{aquacrop_morris}
\alias{aquacrop_morris}
\title{A wrapper function for running a sensitivity analysis for AquaCrop using the morris method}
\usage{
aquacrop_morris(
  situation = "S_01",
  backgroundpar = Spinach,
  r = r,
  binf = c(),
  bsup = c(),
  design = list(type = "oat", levels = 8, grid.jump = 1),
  outvars = c("Biomass")
)
}
\arguments{
\item{situation}{character string or vector of scenarios to run}

\item{backgroundpar}{the reference parameter set (as made by \code{read_CRO})}

\item{r}{number of trajectories in the morris algorithm}

\item{binf}{named vector with lower boundaries of the parameter ranges}

\item{bsup}{named vector with upper boundaries of the parameter ranges}

\item{design}{list for designing the morris algorithm}

\item{outvars}{character vector with output variable names for the sensitivity analysis}
}
\value{
a list of class \code{morris}, containing all the input arguments, plus the following components:

\code{call}: the matched call

\code{X} a \code{data.frame} containing the design of experiments (parameter value combinations)

\code{y} either a vector, a matrix or a three-dimensional array of model responses (depends on the output of model)

\code{ee}:
\itemize{
\item if y is a vector: a \eqn{(r \times p)} - matrix of elementary effects for all the factors.
\item if y is a matrix: a \eqn{(r \times p \times ncol(y))} - array of elementary effects for all the factors and all columns of y.
\item if y is a three-dimensional array: a \eqn{(r \times p \times dim(y)[2] \times dim(y)[3])} - array of elementary effects for all the factors and all elements of the second and third dimension of y.
}
}
\description{
\code{aquacrop_morris} operates in 5 steps:
\enumerate{
\item Define the morris design in terms of parameters, parameter ranges, trajectories and method. See also the \code{morris} function in the \url{sensitivity package}.
\item Run the AquaCrop model with all parameter combinations from the morris design.
\item Choose the level of integration: model time steps, different variables
\item Make an array with the correct dimensions based on the simulations (step 2) and the integration level (step 3).
\item \code{tell} the array to the morris design from step 1.
}
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
Scenario_s <- design_scenario(name = "S_01",
    Plant_Date = as.Date("2019-04-01"),
    IRRI = "IRRI_01",
    Soil = "Soil_01",
    Plu = "Plu_01",
    Tnx = "Tnx_01",
    ETo = "ETo_01")

p <- aquacrop_morris(situation = "S_01", backgroundpar=Spinach, r = 20, binf=c(rt_max = 0.12, cgc = 0.1), bsup = c(rt_max = 0.55, cgc = 0.21), outvars = c("Biomass", "YSdryS"))

}
