% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquacrop_wrapper.R
\name{aquacrop_wrapper}
\alias{aquacrop_wrapper}
\title{A wrapper function for running AquaCrop for use in CroptimizR package}
\usage{
aquacrop_wrapper(param_values, situation, model_options, ...)
}
\arguments{
\item{param_values}{list of crop parameters to modify from the default}

\item{situation}{character vector of scenario names as listed in the Scenario_s tibble}

\item{model_options}{list of model options}
}
\value{
a dataframe of daily outputs of AquaCrop, concatenated over the different scenario's.
}
\description{
\code{aquacrop_wrapper} takes parameter values as an input for an AquaCrop simulation and returns a result list as requested by the
\code{CroptimizR} package.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
