% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AquaCropR-package.R
\docType{package}
\name{AquacropOnR-package}
\alias{AquacropOnR}
\alias{AquacropOnR-package}
\title{AquacropOnR: Function collection for using AquaCrop from R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package provides wrappers for using AquaCrop from within R as simple as possible. This work was inspired on earlier initiatives for the APSIM (ApsimOnR) and STICS (SticsOnR) crop models. In doing so, the package aims to intercface AquaCRop to the CroptimizR package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tdeswaef.github.io/aquacroponr/}
}

}
\author{
\strong{Maintainer}: Tom De Swaef \email{tom.deswaef@ilvo.vlaanderen.be} (\href{https://orcid.org/0000-0002-8338-7786}{ORCID})

}
\keyword{internal}
