% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AquaCropR-package.R
\docType{package}
\name{AquaCropR-package}
\alias{AquaCropR}
\alias{AquaCropR-package}
\title{AquaCropR: Function collection for using AquaCrop from R}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Tom De Swaef \email{tom.deswaef@ilvo.vlaanderen.be} (\href{https://orcid.org/0000-0002-8338-7786}{ORCID})

}
\keyword{internal}
