% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_FileSetup.R
\name{path_config}
\alias{path_config}
\title{Setup of AquaCrop path}
\usage{
path_config(AquaCrop.path)
}
\arguments{
\item{AquaCrop.path}{character name of the folder where the aquacrop.exe file is located}
}
\value{
a character name of the path of the aquacrop.exe file, but also writes the \emph{DailyResults.SIM} file in the appropriate folder. file
}
\description{
Returns the correct the AquaCrop.exe path, while checking the required folder structure. Be aware that the
working directory should be set to the the folder where the aquacrop.exe file is located (\code{setwd()}).
}
