% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_CRO.R
\name{read_CRO}
\alias{read_CRO}
\title{Read a .CRO file and return it as a list}
\usage{
read_CRO(default_crop)
}
\arguments{
\item{default_crop}{file path of the .CRO file to be read.}
}
\value{
a named list with crop parameter name-value pairs.
}
\description{
Read a .CRO file and return it as a list
}
\details{
The crop file required for AquaCrop 7.0 and higher requires 83 lines of content.
Crop files generated using previous AquaCrop versions, typically contain less lines (76).
The `read_CRO` function detects whether the file has sufficient length, and complements the file if needed.
}
