% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_AquaCrop.R
\name{solve_AquaCrop}
\alias{solve_AquaCrop}
\title{A wrapper function for running AquaCrop}
\usage{
solve_AquaCrop(croppar, defaultpar, scenario_s)
}
\arguments{
\item{croppar}{list of crop parameters to modify from the default}

\item{defaultpar}{list of default parameter values}

\item{scenario_s}{character vector of scenario names as listed in the Scenario_s tibble}
}
\value{
a dataframe of daily outputs of AquaCrop, concatenated over the different scenario's.
}
\description{
\code{solve_AquaCrop} takes parameter values as an input for an AquaCrop simulation and returns a dataframe as an output
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
