% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquacrop_season.R
\name{write_scenario_files}
\alias{write_scenario_files}
\title{A function for writing the AquaCrop files}
\usage{
write_scenario_files(
  situation,
  output_vars = c(1, 2),
  cycle_length = 70,
  growth_length = NULL,
  ...
)
}
\arguments{
\item{situation}{character vector of scenario names as listed in the Scenario_s tibble}

\item{output_vars}{defines the output variables that are generated by AquaCrop}

\item{cycle_length}{an integer representing the length of the \strong{simulation} period in days,}

\item{growth_length}{an integer representing the length of the \strong{growth} period in days,}
}
\description{
\code{write_scenario_files} takes scenarios and settings as inputs and writes the aquacrop files.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
