% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquacrop_wrapper.R
\name{aquacrop_wrapper}
\alias{aquacrop_wrapper}
\title{A wrapper function for running AquaCrop for use in CroptimizR package}
\usage{
aquacrop_wrapper(
  param_values,
  situation,
  model_options = list(AQ = AQ, cycle_length = 70, defaultpar = Spinach, output =
    "croptimizr", daily_output = c(1, 2)),
  ...
)
}
\arguments{
\item{param_values}{list of crop parameters to modify from the default}

\item{situation}{character vector of scenario names as listed in the Scenario_s tibble}

\item{model_options}{list of model options: \code{AQ} is the aquacrop path as generated by the \code{path_config} function,
\code{cycle_length} an integer representing the length of the \strong{simulation} period in days, \code{defaultpar} is the default parameter list as designed by the \code{read_CRO} function,
\code{output} defines the shape of the output that is returned and \code{daily_output} defines the daily output variables that are generated by AquaCrop.}
}
\value{
a dataframe or list (if \code{model_options$output} = 'croptimizr') of daily outputs of AquaCrop, concatenated over the different scenario's.
}
\description{
\code{aquacrop_wrapper} takes parameter values as an input for an AquaCrop simulation and returns a result list as requested by the
\code{CroptimizR} package.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{

y <- aquacrop_wrapper(param_values = params,
                      situation = Scenario_s$Scenario,
                      model_options = list(AQ=AQ, defaultpar=Spinach,
                                           output = "croptimizr", daily_output = c(1,2)))

}
