% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_scenario.R
\name{design_scenario}
\alias{design_scenario}
\title{Makes the required tibble that is used as scenario argument in the `solve_AquaCrop()` function.}
\usage{
design_scenario(name, Plant_Date, IRRI, Soil, Plu, Tnx, ETo)
}
\arguments{
\item{name}{character vector with names of the scenarios.}

\item{Plant_Date}{Date format vector with planting dates for the different scenarios}

\item{IRRI}{character vector with names of Irrigation tibble(s)}

\item{Soil}{character vector with names of Soil tibble(s)}

\item{Plu}{character vector with names of Precipitation tibble(s)}

\item{Tnx}{character vector with names of Temperature tibble(s)}

\item{ETo}{character vector with names of Reference Evapotranspiration tibble(s)}
}
\value{
A tibble with the different scenarios to run in AquaCrop
}
\description{
Makes the required tibble that is used as scenario argument in the `solve_AquaCrop()` function.
}
