% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_scenario.R
\name{design_scenario}
\alias{design_scenario}
\title{Makes the required tibble that is used as scenario argument in the \code{aquacrop_wrapper} function.}
\usage{
design_scenario(
  name,
  Input_Date,
  Sim_Date = NULL,
  Plant_Date,
  IRRI,
  Soil,
  Plu,
  Tnx,
  ETo,
  FMAN = "default",
  GWT = 2,
  SW0 = "FC"
)
}
\arguments{
\item{name}{character vector with names of the scenarios.}

\item{Input_Date}{Date format vector with initial input dates for the different scenarios}

\item{Sim_Date}{Date format vector with dates for starting simulation. if NULL, set to Plant_Date}

\item{Plant_Date}{Date format vector with planting dates for the different scenarios}

\item{IRRI}{character vector with names for the ID in the IRRI_s tibble}

\item{Soil}{character vector with names of Soil tibble(s)}

\item{Plu}{character vector with names of Precipitation tibble(s)}

\item{Tnx}{character vector with names of Temperature tibble(s)}

\item{ETo}{character vector with names of Reference Evapotranspiration tibble(s)}

\item{FMAN}{character vector with names for the ID in the FMAN_s tibble}

\item{GWT}{numeric vector with depth of ground water table in m}

\item{SW0}{character vector with names for the ID in the SW0_s tibble. If "FC" initial conditions are taken from SOL_s tibble.}
}
\value{
A tibble with the different scenarios to run in AquaCrop
}
\description{
Makes the required tibble that is used as scenario argument in the \code{aquacrop_wrapper} function.
}
