% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquacrop_season.R
\name{run_aquacrop_season}
\alias{run_aquacrop_season}
\title{A wrapper function for running AquaCrop and only keep seasonal results}
\usage{
run_aquacrop_season(param_values, AQ, defaultpar, cycle_length = cycle_length)
}
\arguments{
\item{param_values}{list of crop parameters to modify from the default}

\item{AQ}{is the aquacrop path as generated by the \code{path_config} function,}

\item{defaultpar}{is the default parameter list as designed by the \code{read_CRO} function,}

\item{cycle_length}{an integer representing the length of the \strong{simulation} period in days,}

\item{situation}{character vector of scenario names as listed in the Scenario_s tibble}
}
\value{
a dataframe of seasonal outputs of AquaCrop, concatenated over the different scenario's.
}
\description{
\code{run_aquacrop_season} takes parameter values as an input for an AquaCrop simulation and returns seasonal values.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
